test: split test_worktree_hooks.py into per-module files (PS-204)
